Refactor emit phase-1 setup
